requirements.txt
aiogram==3.10.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic-settings==2.5.2
pydantic==2.9.2
python-dotenv==1.0.0
pandas==2.2.0
openpyxl==3.1.2
